Document the need to reference fixture files
In order for a fixture file to populate the Database(aka be usable) in a test class you must set the `fixture_file` property of that class to your `YAML` file. That was not mentioned in these docs before so I'm including it. Feel free to suggest copy changes as I'm not sure how robust the explanation should be.